genre

+++ src/main/java/com/springboot/moov/data/dto/GenreDto.java
@@ -11,12 +11,13 @@ public class GenreDto {
     private String title;
     private String genre;
     private String posterUrl;
-    private Double rating;
+    private float rating; // float 타입으로 유지
     private LocalDate releaseDate;
 
     public GenreDto() {}
 
-    public GenreDto(Long id, String title, String genre, String posterUrl, Double rating, LocalDate releaseDate) {
+    // 생성자에서 rating을 float으로 처리
+    public GenreDto(Long id, String title, String genre, String posterUrl, float rating, LocalDate releaseDate) {
         this.id = id;
         this.title = title;
         this.genre = genre;
